+++ src/Events/Authentication/SuccessfullyLoggedIn.ts
@@ -4,8 +4,8 @@ import {
   AuthContextModule,
   AuthIntent,
   AuthService,
+  SuccessfullyLoggedIn,
 } from "../../Schema"
-import { SuccessfullyLoggedIn } from "../../Schema/Event"
 
 export interface SuccessfullyLoggedInArgs {
   authRedirect: string
@@ -26,7 +26,7 @@ export interface SuccessfullyLoggedInArgs {
  *   authRedirect: "https://artsy.net/artist/andy-warhol",
  *   contextModule: ContextModule.popUpModal,
  *   copy: "Sign up to follow artists",
- *   intent: AuthIntent.viewArtist,
+ *   intent: Intent.viewArtist,
  *   service: "email",
  *   triggerSeconds: 3,
  *   userId: "5bd8b675776bd6002c86526c"
