--- src/Schema/ActionType.ts
@@ -1,21 +1,0 @@
-/**
- * The top-level actions an Event describes.
- *
- * Each ActionType corresponds with a table in Redshift.
- * @packageDocumentation
- */
-export enum ActionType {
-  /**
-   * Corresponds to {@link AuthImpression}
-   */
-  authImpression = "authImpression",
-  createdAccount = "createdAccount",
-  resetYourPassword = "resetYourPassword",
-  successfullyLoggedIn = "successfullyLoggedIn",
-  tappedArtistGroup = "tappedArtistGroup",
-  tappedArtworkGroup = "tappedArtworkGroup",
-  tappedAuctionGroup = "tappedAuctionGroup",
-  tappedCollectionGroup = "tappedCollectionGroup",
-  tappedExploreGroup = "tappedExploreGroup",
-  tappedFairGroup = "tappedFairGroup"
-}
