+++ src/Schema/ContextModule.ts
@@ -1,4 +1,5 @@
 /**
+ * Master list of context modules available in web and iOS
  * A component where an action is triggered
  * @packageDocumentation
  */
@@ -62,5 +63,54 @@ export enum ContextModule {
   tagHeader = "tagHeader",
   trendingArtistsRail = "trendingArtistsRail",
   worksByPopularArtistsRail = "worksByPopularArtistsRail",
-  worksForSaleRail = "worksForSaleRail"
+  worksForSaleRail = "worksForSaleRail",
 }
+
+/**
+ * Limited ContextModules available for web authentication events
+ */
+export type AuthContextModule =
+  | ContextModule.aboutTheWork
+  | ContextModule.artistHeader
+  | ContextModule.artistRecentlySold
+  | ContextModule.artistsTab
+  | ContextModule.artistsToFollowRail
+  | ContextModule.artworkGrid
+  | ContextModule.artworkImage
+  | ContextModule.artworkSidebar
+  | ContextModule.auctionSidebar
+  | ContextModule.auctionRail
+  | ContextModule.auctionResults
+  | ContextModule.auctionsInfo
+  | ContextModule.bannerPopUp
+  | ContextModule.browseFair
+  | ContextModule.categoryRail
+  | ContextModule.collectionDescription
+  | ContextModule.consignSubmissionFlow
+  | ContextModule.currentShowsRail
+  | ContextModule.fairInfo
+  | ContextModule.fairsHeader
+  | ContextModule.featuredArtistsRail
+  | ContextModule.footer
+  | ContextModule.geneHeader
+  | ContextModule.header
+  | ContextModule.intextTooltip
+  | ContextModule.liveAuctionsRail
+  | ContextModule.mainCarousel
+  | ContextModule.minimalCTABanner
+  | ContextModule.otherWorksByArtistRail
+  | ContextModule.otherWorksFromPartnerRail
+  | ContextModule.otherWorksFromShowRail
+  | ContextModule.otherWorksInAuctionRail
+  | ContextModule.partnerHeader
+  | ContextModule.pastFairs
+  | ContextModule.popUpModal
+  | ContextModule.recentlyViewedRail
+  | ContextModule.relatedArtistsRail
+  | ContextModule.relatedWorksRail
+  | ContextModule.saveWorksCTA
+  | ContextModule.showHeader
+  | ContextModule.showInfo
+  | ContextModule.tagHeader
+  | ContextModule.worksByPopularArtistsRail
+  | ContextModule.worksForSaleRail
