+++ src/Schema/index.ts
@@ -1,4 +1,6 @@
-export * from "./ActionType"
-export * from "./Authentication"
 export * from "./ContextModule"
 export * from "./Event"
+export * from "./Events/Authentication"
+export * from "./Events/Tap"
+export * from "./Intent"
+export * from "./OwnerType"
